Implement new Parser class and integrate with existing code
Changed the Command interface to an abstract class so as to facilitate the
sharing of common members such as INDEX_FIELD, INDEX_DATA and the
main internships list. Also, added a function so as to set from the main loop -
the common internship list for all commands to access, read and modify,

Deleted constructors within the subclass commands as they are instantiated
from the parent abstract class already. Changed the parameter type of the execute
function from String[] to ArrayList<String> so that we can call more functions
to easily manipulate the arguments that are passed into the function.

Gave a default id value for internships which haven't been instantiated yet.

Added a new file under the test folder just to showcase
how to operate the program and the expected behaviour/syntax.
Sample inputs and outputs only cover the basic cases without testing
too many edge cases.