Fix wrong numbering of see all incomes command
See all incomes command prints the index of each income relative to
all financial entries instead of relative to recorded incomes only.

Let's add a private variable incomeCount which records the number of
incomes printed so far. incomeCount is printed together with each
income, serving as the correct index of the recorded income.