Replace InvalidCommand with an exception
User command errors are handled by an InvalidCommand, which then passes
the error message back to the user. This makes the code less readable,
as it is not obvious that InvalidCommand handles errors.

Let's replace InvalidCommand with an exception, that is handled by the
UI. This way, the code is easier to read as it is clearer how the
user command errors are handled.