Add error handling for adding participants
When participants are added to events that are not present in the event
list, the program informs the user that the participant was added,
although neither the participant nor the event are stored in the
program. This is misleading to the user.

Let's amend the add participant command to show the user a message that
the adding failed for this case. This makes the add command more
inituitive to the user.