Change message syntax for participant details
The participants' status is denoted using the terms "present", if they
are present, and "not present", if they are not present. This makes the
participant list entries long and difficult to read.

Let's change the syntax such that whether a participant has been marked
present or not is indicated by the presence or absence of an "X"
beside the participant's name. This makes the participant list easier
to read.