Amend addEntry method to insert entry according to order of date
App stores and lists entries in order of time of entering event.
It would be more user-friendly if it stored and listed all entries
by ascending order of date of entry.

Let's amend the add entry function to always insert entries by order of
date.

We continue to use an ArrayList for the financial list to minimise
changes required in other parts of the code. The method checks whether
to insert entries from the end of the ArrayList as it is more likely
for users to add entries in ascending order of date, thus this should
reduce the average time needed to find the correct index to insert the
Financial entry.