Fix List Entries commands to include entries on start/end date
App does not include entries which have the same date as the start/end
date specified when calling the list command. This is not intuitive
for the user.

Let's amend the commands such that when listing the financial entries,
entries which fall on the entered start/end date will also be listed.

Furthermore, the conditional to check if an entry falls within the
specified date range was extracted to a new method in
SeeAllEntriesCommands to improve code readability. Furthermore, this
method was set to protected to allow inheritance by
SeeAllExpensesCommand and SeeAllIncomesCommand, thus reducing
repeated code.