Change where checkLimitOf method is used.
Before, it was called when addCategory is used, and also editCategory
(because editCategory calls addCategory).

Now, that method is called in the Command class instead so that when
loading from csv file the messages are not spammed (since loadFromCsv
uses addCategory).

To do so, visibility of checkLimitOf is changed from private to public