Edit EditEntryCommand to fix bug where edit changes order unnecessarily
Edit command currently deletes the entry at the specified index and
adds a replacement entry to the list. This leads to the order of the
entry changing in the list even if there is no change in date, which
could throw users off.

Let's:
* Modify EditEntryCommand to increment/decrement the index of the
entry from its original index until it is correctly sorted in the
Financial List.

This prevents unnecessary changes in order of entries in the list.