Fix bug where exception thrown when editing category of an expense
This occurs when editing category of expense, and there are no more
expenses with the old category.

Before, the expense's category will get updated to newCategory.
Tracker is then edited, and attempts to decrement total of oldCategory.
To do so, it needs to call getCategoryDataOf first to find oldCategory.
However, key of oldCategory is no longer in the tracker because the
category of expense is newExpense and no longer oldExpense.
This causes getCategoryDataOf throws an exception.